Move eval APIs into excluded modules
